Contains several fixes for docs/ref/class-based-views, as follows:
Restored the section headings which duplicate the class directives (removed in [14264]) so that a full table of contents is still generated. Thanks to Jacob for the suggestion.

Fixed a handful of typos, and corrected a method directive that was mistakenly labeled as an attribute.

Expanded method descriptions for SingleObjectMixin to better explain the default behaviors.

Reformatted ProcessFormView docs to be consistent.

Added missing description of MultipleObjectMixin.get_paginate_by.

--HG--
extra : convert_revision : svn%3Abcc190cf-cafb-0310-a4f2-bffc1f526a37/django/trunk%4014266